FIX: signed/unsigned comparison warnings
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@4660 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/apps/gmm.cxx
@@ -195,7 +195,7 @@ doMain
     cmtk::StdOut << "\n";
 
 #pragma omp parallel for    
-    for ( int n = 0; n < nPixels; ++n )
+    for ( int n = 0; n < static_cast<int>( nPixels ); ++n )
       {
       if ( maskImage->GetDataAt( n ) > 0 )
 	{
@@ -243,7 +243,7 @@ doMain
     }
   
 #pragma omp parallel for
-  for ( int n = 0; n < nPixels; ++n )
+  for ( int n = 0; n < static_cast<int>( nPixels ); ++n )
     {
     if ( maskImage->GetDataAt( n ) > 0 )
       {
